Opt-in keep test-compile before package (#1)
#2 There is an interesting question, which order of phases is better.
Latest `0.7` version has order `[compile, package, test-compile, test, integration-test, verify, install, deploy]`: `prepare-package` and `package` moved before `test-compile` and `test` (some phases omitted to be concise).

It has an implicit side effect: if the command is `mvn package`, now the test compilation is skipped. This can be fine, but can confuse as well. This can be solved by changing the command to `mvn test-compile` or `mvn install` which is also confusing and also usually not recommended. The problem is that phases are patched only when `-b turbo` parameter is added, so these changes should be done in combination (enable builder and change the requested phase).

This PR is opt-in with phase order `[compile, test-compile, package, test..]`: `package` is before `test`, but after `test-compile`. Note that it's not the same as current impl `package` before `test-compile`. This will reduce the optimization effect a bit (especially if there is only a compilation), but if the elephant in the room is running surefire/failsafe, that's still a huge boost.

<img width="721" alt="Screenshot 2025-05-29 at 13 20 15" src="https://github.com/user-attachments/assets/8309a90b-d22f-4a46-8a10-13b3335661f4" />

To change the order of phases, specify the parameter `-DturboTestCompile` in the command line arguments or in `.mvn/maven.config`.

This also allows usage of `test-jar` packaging if needed.